requirements.txt
flask>=3.0.0
flask-session>=0.5.0
bcrypt>=4.1.0
werkzeug>=3.0.0
numpy>=1.24.0
h5py>=3.9.0
matplotlib>=3.7.0
Pillow>=10.0.0
# tensorflow>=2.13.0   # uncomment when deploying with GPU/CPU model